+++ packages/iota/examples/shimmer-value-transaction-tutorial/src/generate-addresses.ts
@@ -1,0 +1,55 @@
+import { Bip32Path, Bip39 } from "@iota/crypto.js";
+import { Bech32Helper, Ed25519Address, Ed25519Seed, ED25519_ADDRESS_TYPE, generateBip44Address, IKeyPair } from "@iota/iota.js";
+import { Converter } from "@iota/util.js";
+
+// Default entropy length is 256
+const randomMnemonic = Bip39.randomMnemonic();
+
+console.log("Seed phrase:", randomMnemonic);
+
+const masterSeed = Ed25519Seed.fromMnemonic(randomMnemonic);
+
+const NUM_ADDR = 6;
+const addressGeneratorAccountState = {
+    accountIndex: 0,
+    addressIndex: 0,
+    isInternal: false
+};
+const paths: string[] = [];
+for (let i = 0; i < NUM_ADDR; i++) {
+    const path = generateBip44Address(addressGeneratorAccountState);
+    paths.push(path);
+
+    console.log(`${path}`);
+}
+
+
+const keyPairs: IKeyPair[] = [];
+
+for (const path of paths) {
+    // Master seed was generated previously
+    const addressSeed = masterSeed.generateSeedFromPath(new Bip32Path(path));
+    const addressKeyPair = addressSeed.keyPair();
+    keyPairs.push(addressKeyPair);
+
+    console.log(Converter.bytesToHex(addressKeyPair.privateKey, true));
+    console.log(Converter.bytesToHex(addressKeyPair.publicKey, true));
+}
+
+const publicAddresses: { ed25519: string, bech32: string }[] = [];
+
+for (const keyPair of keyPairs) {
+    const ed25519Address = new Ed25519Address(keyPair.publicKey);
+    // Address in bytes
+    const ed25519AddressBytes = ed25519Address.toAddress();
+    // Conversion to BECH32
+    const bech32Addr = Bech32Helper.toBech32(ED25519_ADDRESS_TYPE, ed25519AddressBytes, "rms");
+
+    const publicAddress = {
+        ed25519: Converter.bytesToHex(ed25519AddressBytes, true),
+        bech32: bech32Addr
+    };
+    publicAddresses.push(publicAddress);
+
+    console.log(publicAddress);
+}
