+++ packages/iota/examples/shimmer-value-transaction-tutorial/src/send-value-transactions.ts
@@ -1,17 +1,12 @@
-import {Bip32Path, Bip39, Blake2b, Ed25519} from "@iota/crypto.js";
+import { Blake2b, Ed25519} from "@iota/crypto.js";
 import {
     ADDRESS_UNLOCK_CONDITION_TYPE,
     BASIC_OUTPUT_TYPE,
-    Bech32Helper,
     DEFAULT_PROTOCOL_VERSION,
     ED25519_ADDRESS_TYPE,
     ED25519_SIGNATURE_TYPE,
-    Ed25519Address,
-    Ed25519Seed,
-    generateBip44Address,
     IBasicOutput,
     IBlock,
-    IKeyPair,
     ISignatureUnlock,
     ITransactionEssence,
     ITransactionPayload,
@@ -26,58 +21,7 @@ import {
 import {Converter, WriteStream} from "@iota/util.js";
 import {NeonPowProvider} from "@iota/pow-neon.js";
 import bigInt from "big-integer";
-
-// Default entropy length is 256
-const randomMnemonic = Bip39.randomMnemonic();
-
-console.log("Seed phrase:", randomMnemonic);
-
-const masterSeed = Ed25519Seed.fromMnemonic(randomMnemonic);
-
-const NUM_ADDR = 6;
-const addressGeneratorAccountState = {
-    accountIndex: 0,
-    addressIndex: 0,
-    isInternal: false
-};
-const paths: string[] = [];
-for (let i = 0; i < NUM_ADDR; i++) {
-    const path = generateBip44Address(addressGeneratorAccountState);
-    paths.push(path);
-
-    console.log(`${path}`);
-}
-
-
-const keyPairs: IKeyPair[] = [];
-
-for (const path of paths) {
-    // Master seed was generated previously
-    const addressSeed = masterSeed.generateSeedFromPath(new Bip32Path(path));
-    const addressKeyPair = addressSeed.keyPair();
-    keyPairs.push(addressKeyPair);
-
-    console.log(Converter.bytesToHex(addressKeyPair.privateKey, true));
-    console.log(Converter.bytesToHex(addressKeyPair.publicKey, true));
-}
-
-const publicAddresses: { ed25519: string, bech32: string }[] = [];
-
-for (const keyPair of keyPairs) {
-    const ed25519Address = new Ed25519Address(keyPair.publicKey);
-    // Address in bytes
-    const ed25519AddressBytes = ed25519Address.toAddress();
-    // Conversion to BECH32
-    const bech32Addr = Bech32Helper.toBech32(ED25519_ADDRESS_TYPE, ed25519AddressBytes, "rms");
-
-    const publicAddress = {
-        ed25519: Converter.bytesToHex(ed25519AddressBytes, true),
-        bech32: bech32Addr
-    };
-    publicAddresses.push(publicAddress);
-
-    console.log(publicAddress);
-}
+import { exit } from "process";
 
 const API_ENDPOINT = "https://api.testnet.shimmer.network";
 const client = new SingleNodeClient(API_ENDPOINT, {powProvider: new NeonPowProvider()});
@@ -92,8 +36,13 @@ const sourceAddressPrivateKey = "0x87b145379dbdf681fec39b026a6a1f3a40ad9ce18f523
 
 const destAddress = "0xf8cec8dca90daf5498c3c2f5db880c1fc759bd0e4cf40b0e12c297fde3c731a1";
 
-const consumedOutputId = "0xe5b4dff0c6c5d49de373cb936894c85f52b46a05972535fbf2910be3e8eb4ebf0100";
+const consumedOutputId = "0x8642289fd18e94af55643bdf91850b79df9c90389fef0d2b9f2bbe4d1f35ff810100";
 const outputDetails = await client.output(consumedOutputId);
+if (outputDetails.metadata.isSpent) {
+    console.error(`Output ${consumedOutputId} is already spent. Use another output controlled by ${sourceAddressBech32}`);
+    exit(-1);
+}
+
 const totalFunds = bigInt(outputDetails.output.amount);
 
 const amountToSend = bigInt("50000");
