+++ packages/iota/examples/shimmer-value-transaction-tutorial/src/sweep-outputs.ts
@@ -1,17 +1,12 @@
-import {Bip32Path, Bip39, Blake2b, Ed25519} from "@iota/crypto.js";
 import {
     ADDRESS_UNLOCK_CONDITION_TYPE,
     BASIC_OUTPUT_TYPE,
-    Bech32Helper,
     DEFAULT_PROTOCOL_VERSION,
     ED25519_ADDRESS_TYPE,
     ED25519_SIGNATURE_TYPE,
-    Ed25519Address,
-    Ed25519Seed,
-    generateBip44Address,
     IBasicOutput,
     IBlock,
-    IKeyPair, IndexerPluginClient, IReferenceUnlock,
+    IndexerPluginClient, IReferenceUnlock,
     ISignatureUnlock,
     ITransactionEssence,
     ITransactionPayload,
@@ -26,57 +21,8 @@ import {
 import {Converter, WriteStream} from "@iota/util.js";
 import {NeonPowProvider} from "@iota/pow-neon.js";
 import bigInt from "big-integer";
-
-// Default entropy length is 256
-const randomMnemonic = Bip39.randomMnemonic();
-
-console.log("Seed phrase:", randomMnemonic);
-
-const masterSeed = Ed25519Seed.fromMnemonic(randomMnemonic);
-
-const NUM_ADDR = 6;
-const addressGeneratorAccountState = {
-    accountIndex: 0,
-    addressIndex: 0,
-    isInternal: false
-};
-const paths: string[] = [];
-for (let i = 0; i < NUM_ADDR; i++) {
-    const path = generateBip44Address(addressGeneratorAccountState);
-    paths.push(path);
-
-    console.log(`${path}`);
-}
-
-const keyPairs: IKeyPair[] = [];
-
-for (const path of paths) {
-    // Master seed was generated previously
-    const addressSeed = masterSeed.generateSeedFromPath(new Bip32Path(path));
-    const addressKeyPair = addressSeed.keyPair();
-    keyPairs.push(addressKeyPair);
-
-    console.log(Converter.bytesToHex(addressKeyPair.privateKey, true));
-    console.log(Converter.bytesToHex(addressKeyPair.publicKey, true));
-}
-
-const publicAddresses: { ed25519: string, bech32: string }[] = [];
-
-for (const keyPair of keyPairs) {
-    const ed25519Address = new Ed25519Address(keyPair.publicKey);
-    // Address in bytes
-    const ed25519AddressBytes = ed25519Address.toAddress();
-    // Conversion to BECH32
-    const bech32Addr = Bech32Helper.toBech32(ED25519_ADDRESS_TYPE, ed25519AddressBytes, "rms");
-
-    const publicAddress = {
-        ed25519: Converter.bytesToHex(ed25519AddressBytes, true),
-        bech32: bech32Addr
-    };
-    publicAddresses.push(publicAddress);
-
-    console.log(publicAddress);
-}
+import { Blake2b, Ed25519 } from "@iota/crypto.js";
+import { exit } from "process";
 
 const API_ENDPOINT = "https://api.testnet.shimmer.network";
 const client = new SingleNodeClient(API_ENDPOINT, {powProvider: new NeonPowProvider()});
@@ -91,6 +37,11 @@ const indexerPlugin = new IndexerPluginClient(client);
 const outputList = await indexerPlugin.basicOutputs({
     addressBech32: destAddressBech32
 });
+
+if (outputList.items.length < 2) {
+    console.error(`To sweep funds you need at least 2 outputs controlled by address ${destAddressBech32}`);
+    exit(-1);
+}
 
 const consumedOutputId1 = outputList.items[0];
 const consumedOutputId2 = outputList.items[1];
